editors - adopt working copy handler for files
