+++ src/vs/workbench/contrib/files/browser/files.ts
@@ -3,10 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { URI, UriComponents } from 'vs/base/common/uri';
+import { URI } from 'vs/base/common/uri';
 import { IListService } from 'vs/platform/list/browser/listService';
 import { OpenEditor, ISortOrderConfiguration } from 'vs/workbench/contrib/files/common/files';
-import { EditorResourceAccessor, SideBySideEditor, IEditorIdentifier, EditorInput, IEditorInputSerializer } from 'vs/workbench/common/editor';
+import { EditorResourceAccessor, SideBySideEditor, IEditorIdentifier } from 'vs/workbench/common/editor';
 import { List } from 'vs/base/browser/ui/list/listWidget';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { ExplorerItem } from 'vs/workbench/contrib/files/common/explorerModel';
@@ -14,62 +14,9 @@ import { coalesce } from 'vs/base/common/arrays';
 import { AsyncDataTree } from 'vs/base/browser/ui/tree/asyncDataTree';
 import { IEditorGroupsService } from 'vs/workbench/services/editor/common/editorGroupsService';
 import { IEditableData } from 'vs/workbench/common/views';
-import { createDecorator, IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { ResourceFileEdit } from 'vs/editor/browser/services/bulkEditService';
 import { ProgressLocation } from 'vs/platform/progress/common/progress';
-import { FileEditorInput } from 'vs/workbench/contrib/files/common/editors/fileEditorInput';
-import { isEqual } from 'vs/base/common/resources';
-
-interface ISerializedFileEditorInput {
-	resourceJSON: UriComponents;
-	preferredResourceJSON?: UriComponents;
-	name?: string;
-	description?: string;
-	encoding?: string;
-	modeId?: string;
-}
-
-export class FileEditorInputSerializer implements IEditorInputSerializer {
-
-	canSerialize(editorInput: EditorInput): boolean {
-		return true;
-	}
-
-	serialize(editorInput: EditorInput): string {
-		const fileEditorInput = editorInput as FileEditorInput;
-		const resource = fileEditorInput.resource;
-		const preferredResource = fileEditorInput.preferredResource;
-		const serializedFileEditorInput: ISerializedFileEditorInput = {
-			resourceJSON: resource.toJSON(),
-			preferredResourceJSON: isEqual(resource, preferredResource) ? undefined : preferredResource, // only storing preferredResource if it differs from the resource
-			name: fileEditorInput.getPreferredName(),
-			description: fileEditorInput.getPreferredDescription(),
-			encoding: fileEditorInput.getEncoding(),
-			modeId: fileEditorInput.getPreferredMode() // only using the preferred user associated mode here if available to not store redundant data
-		};
-
-		return JSON.stringify(serializedFileEditorInput);
-	}
-
-	deserialize(instantiationService: IInstantiationService, serializedEditorInput: string): FileEditorInput {
-		return instantiationService.invokeFunction(accessor => {
-			const serializedFileEditorInput: ISerializedFileEditorInput = JSON.parse(serializedEditorInput);
-			const resource = URI.revive(serializedFileEditorInput.resourceJSON);
-			const preferredResource = URI.revive(serializedFileEditorInput.preferredResourceJSON);
-			const name = serializedFileEditorInput.name;
-			const description = serializedFileEditorInput.description;
-			const encoding = serializedFileEditorInput.encoding;
-			const mode = serializedFileEditorInput.modeId;
-
-			const fileEditorInput = accessor.get(IEditorService).createEditorInput({ resource, label: name, description, encoding, mode, forceFile: true }) as FileEditorInput;
-			if (preferredResource) {
-				fileEditorInput.setPreferredResource(preferredResource);
-			}
-
-			return fileEditorInput;
-		});
-	}
-}
 
 export interface IExplorerService {
 	readonly _serviceBrand: undefined;
