+++ src/vs/workbench/contrib/files/test/browser/fileEditorInput.test.ts
@@ -19,7 +19,7 @@ import { DisposableStore } from 'vs/base/common/lifecycle';
 import { BinaryEditorModel } from 'vs/workbench/common/editor/binaryEditorModel';
 import { IResourceEditorInput } from 'vs/platform/editor/common/editor';
 import { Registry } from 'vs/platform/registry/common/platform';
-import { FileEditorInputSerializer } from 'vs/workbench/contrib/files/browser/files';
+import { FileEditorInputSerializer } from 'vs/workbench/contrib/files/browser/editors/fileEditorHandler';
 
 suite('Files - FileEditorInput', () => {
